Refactoring Source Code - 5 & Modify Editor Function
