+++ frontend/public/index.html
@@ -30,28 +30,5 @@
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
-    <script>
-      function noEvent() {
-        if (event.keyCode == 116) {
-          event.keyCode= 2;
-          return false;
-        }
-        else if(event.ctrlKey && (event.keyCode==78 || event.keyCode == 82))
-        {
-          return false;
-        }
-      }
-      document.onkeydown = noEvent;
-    </script>
   </body>
 </html>
