+++ frontend/src/components/Header.js
@@ -1,7 +1,14 @@
+import styled from "styled-components";
+import { BasicHeaderStructure } from "../utils/CSS";
+
 const Header = () => {
     return (
-        <h1>Dong-A 3D Model WYSIWYG Editor</h1>
+        <HeaderOne>Dong-A 3D Model WYSIWYG Editor</HeaderOne>
     );
 }
 
 export default Header;
+
+const HeaderOne = styled.h1`
+    ${BasicHeaderStructure('4rem')}
+`;
