+++ frontend/src/components/MyPage/MyPageHeader/MyInformation.js
@@ -48,6 +48,7 @@ const MyInformationContainer = styled.div`
       h2 {
         font-family: "Quill";
         margin-top: 0;
+        font-size: 3rem;
         font-weight: bold;
         margin-bottom: 20px;
       }
