+++ frontend/src/components/ThreeDUpload.js
@@ -24,7 +24,7 @@ const ThreeDUpload = ({ threeDName, threeDURL }) => {
       alpha: false, // 배경 투명도
       preserveDrawingBuffer: true,
     });
-    renderer.setSize(1050, 1050);
+    renderer.setSize(window.innerWidth / 2, window.innerHeight / 1.5);
     renderer.setClearColor(0xffffff, 1);
     rendererRef.current = renderer;
 
@@ -157,8 +157,10 @@ const ThreeDUpload = ({ threeDName, threeDURL }) => {
       <div className="threeD-div">
         <canvas ref={canvasRef} />
       </div>
-      <button type="button" onClick={modifySize}>3D 크기 수정하기</button>
-      <button type="button" onClick={downloadThreeD}>3D 파일 다운로드</button>
+      <ButtonContainer>
+        <button type="button" onClick={modifySize}>3D 크기 수정하기</button>
+        <button type="button" onClick={downloadThreeD}>3D 파일 다운로드</button>
+      </ButtonContainer>
     </ThreeDUploadContainer>
   );
 };
@@ -167,21 +169,12 @@ export default ThreeDUpload;
 
 const ThreeDUploadContainer = styled.div`
   max-width: 1200px;
-  margin: 0 auto;
-  padding: 60px;
+  margin-top: 2rem;
+  padding: 2rem;
   align-items: center;
   background-color: white;
   box-shadow: 0 0 20px rgba(0, 0, 0, 0.1);
-  border-radius: 10px;
-
-  h2 {
-    text-align: center;
-    margin-top: 30px;
-    font-size: 4rem;
-    font-weight: 900;
-    color: #fff;
-    text-shadow: 0 1px 0 #ccc, 0 2px 0 #c9c9c9, 0 3px 0 #bbb, 0 4px 0 #b9b9b9, 0 5px 0 #aaa, 0 6px 1px rgba(0, 0, 0, 0.1), 0 0 5px rgba(0, 0, 0, 0.1), 0 1px 3px rgba(0, 0, 0, 0.3), 0 3px 5px rgba(0, 0, 0, 0.2), 0 5px 10px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.2), 0 20px 20px rgba(0, 0, 0, 0.15);
-  }
+  border-radius: 0.5rem;
 
   .threeD-div {
     display: flex;
@@ -191,16 +184,17 @@ const ThreeDUploadContainer = styled.div`
     width: 100%;
 
     canvas {
-      max-width: 100%;
-      height: auto;
       border: 5px solid black;
       border-radius: 10px;
-      background-color: beige;
       padding: 10px;
     }
   }
+`;
 
-  button {
-    margin-top: 20px;
-  }
+const ButtonContainer = styled.div`
+  margin-top: 0.5rem;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
 `;
