--- frontend/src/css/Board.css
@@ -1,43 +1,0 @@
-.Board {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-  
-  .Board-h1 {
-    font-family: 'Playfair Display', serif;
-    font-size: 3.5rem;
-    font-weight: 800;
-    color: #222;
-    text-align: center;
-    margin-top: 3rem;
-    margin-bottom: 2rem;
-    text-shadow: 3px 3px 5px rgba(0, 0, 0, 0.2);
-  }
-
-  .pagination {
-    display: flex;
-    list-style: none;
-    padding: 0;
-    margin-bottom: 150px;
-  }
-  
-  .pagination li {
-    margin: 0 5px;
-  }
-  
-  .pagination li a {
-    text-decoration: none;
-    padding: 8px 12px;
-    border: 1px solid #ddd;
-  }
-  
-  .pagination li a:hover {
-    background-color: #f0f0f0;
-  }
-  
-  .active a {
-    background-color: #007bff;
-    color: white;
-  }
