+++ frontend/src/css/MyEditor.scss
@@ -1,7 +1,3 @@
-
-/*본인이 설정하고 싶은 폰트 import하기*/
-@import url("https://webfontworld.github.io/amore/AritaBuri.css");
-
 @font-face {
   font-family: "GangwonEduHyeonokT_OTFMediumA";
   src: url("https://cdn.jsdelivr.net/gh/projectnoonnu/noonfonts_2201-2@1.0/GangwonEduHyeonokT_OTFMediumA.woff")
@@ -88,30 +84,14 @@
 
 /* 에디터 양식 */
 .ql-editor{ 
-  min-height: 500px !important;
-  max-height: 800px;
-  width: 100%;
-  overflow: hidden;
-  overflow-y: scroll;
-  overflow-x: scroll;
-}
-
-.text-editor {
-    margin-bottom: 150px;
-}
-
-.text-editor > Button {
-  margin-top: 30px;
-}
-
-.threeD-Model-h2 {
-  text-align: center;
-  margin-top: 60px;
-  margin-bottom: 20px;
-  font-size: 2rem;
-  font-weight: 900;
-  color: #fff;
-  text-shadow: 0 1px 0 #ccc, 0 2px 0 #c9c9c9, 0 3px 0 #bbb, 0 4px 0 #b9b9b9, 0 5px 0 #aaa, 0 6px 1px rgba(0,0,0,.1), 0 0 5px rgba(0,0,0,.1), 0 1px 3px rgba(0,0,0,.3), 0 3px 5px rgba(0,0,0,.2), 0 5px 10px rgba(0,0,0,.25), 0 10px 10px rgba(0,0,0,.2), 0 20px 20px rgba(0,0,0,.15);
+  width: 100%; 
+  min-height: 70vh;
+  
+  img {
+    border-radius: 10px !important;
+    border: 3px solid #282c34 !important;  
+    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.7); 
+  }
 }
 
 .quill-title {
